## Description

Nice work with the app! I noticed some low hanging perf improvements that could help make things feel a bit snappier.

Some high level comparisons:

```ruby
# http://127.0.0.1:3000/sessions?only_path=true&starts_at=2024-08-31
# Before:
#   Completed 200 OK in 338ms (Views: 180.5ms | ActiveRecord: 4.3ms (128 queries, 22 cached) | GC: 41.3ms)
# After:
#   Completed 200 OK in 228ms (Views: 66.2ms | ActiveRecord: 1.3ms (27 queries, 3 cached) | GC: 17.0ms)

# http://127.0.0.1:3000/sessions?only_path=true
# Before:
#   Completed 200 OK in 842ms (Views: 678.7ms | ActiveRecord: 11.5ms (246 queries, 40 cached) | GC: 304.7ms)
# After:
#   Completed 200 OK in 318ms (Views: 153.8ms | ActiveRecord: 2.2ms (40 queries, 3 cached) | GC: 29.0ms)
```

## How has this been tested?

Please mark the tests that you ran to verify your changes. If difficult to test, consider providing instructions so reviewers can test.

- [x] Manual testing
- [ ] System tests
- [ ] Unit tests
- [ ] None

## Checklist

- [ ] CI pipeline is passing
- [x] My code follows the conventions of this project
- [x] I have performed a self-review of my code
- [ ] I have commented on my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation (if applicable)
- [ ] I have added seed data to the database (if applicable)